Big changes to todo and calendar
Updated the TodoTaskCategories and Task appearances to be much more compact and prettier. Also changed the positioning of the add task and settings buttons to the topbar instead of the navbar. Some layout changes for the calendar to make it more suitable for mobile and more appealing to the eye. Added a notes page that will be used for **notes**! Mobile formatting in general much better and much more intuitive.